+++ app/src/main/kotlin/no/nav/aap/arenaoppslag/modeller/ArenaSakDetaljertRespons.kt
@@ -1,6 +1,5 @@
 package no.nav.aap.arenaoppslag.modeller
 
-import no.nav.aap.arenaoppslag.database.KvotebrukHendelse
 import java.time.LocalDateTime
 
 @Suppress("MatchingDeclarationName")
