+++ app/src/test/kotlin/no/nav/aap/arenaoppslag/database/TelleverkForPersonRepositoryTest.kt
@@ -1,5 +1,6 @@
 package no.nav.aap.arenaoppslag.database
 
+import no.nav.aap.arenaoppslag.modeller.KvoteVerdi
 import no.nav.aap.arenaoppslag.modeller.PersonId
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
